Correction to previous code

+++ firmware/ebike_app.c
@@ -766,10 +766,10 @@ void read_throotle (void)
   // map throttle value from 0 up to 255
   ui8_throttle_value = (uint8_t) (map (
 		  ui8_adc_throttle_value,
-		  ADC_THROTTLE_MIN_VALUE,
-		  ADC_THROTTLE_MAX_VALUE,
-		  THROTTLE_MIN_VALUE,
-		  THROTTLE_MAX_VALUE));
+		  (uint8_t) ADC_THROTTLE_MIN_VALUE,
+		  (uint8_t) ADC_THROTTLE_MAX_VALUE,
+		  (uint8_t) THROTTLE_MIN_VALUE,
+		  (uint8_t) THROTTLE_MAX_VALUE));
 
   // low pass filter the torque sensor to smooth the signal
   ui16_throttle_value_accumulated -= ui16_throttle_value_accumulated >> 2;
